azure-functions/requirements.txt
# Azure Functions runtime
azure-functions==1.19.0

# Azure IoT Hub — trigger de mensajes + control C2D
azure-iot-hub==2.6.1

# Cosmos DB — almacenamiento de series de tiempo
azure-cosmos==4.6.0

# Umbrales de alerta (variables de entorno, sin dependencias extra)
# alertas.py en shared/ no requiere librerías externas
